+++ src/chat/component/index.ts
@@ -1,3 +1,4 @@
 export { CopyIcon } from "./CopyIcon";
 export { ScrollView } from "./ScrollView";
 export { OPFSImage } from "./OPFSImage";
+export * from "./ToolUsagePopup";